Consolidate the react-native client-side bridge
